Rename the main and auxiliary glimpse source files.
