--- src/glimpse_main.cpp
@@ -1,57 +1,0 @@
-/*! Copyright UCL, 2019
- * author : Tim Spain <t.spain@ucl.ac.uk>
- *
- * This software is a computer program whose purpose is to reconstruct mass maps
- * from weak gravitational lensing.
- *
- * This software is governed by the CeCILL license under French law and
- * abiding by the rules of distribution of free software.  You can  use,
- * modify and/ or redistribute the software under the terms of the CeCILL
- * license as circulated by CEA, CNRS and INRIA at the following URL
- * "http://www.cecill.info".
- *
- * As a counterpart to the access to the source code and  rights to copy,
- * modify and redistribute granted by the license, users are provided only
- * with a limited warranty  and the software's author,  the holder of the
- * economic rights,  and the successive licensors  have only  limited
- * liability.
- *
- * In this respect, the user's attention is drawn to the risks associated
- * with loading,  using,  modifying and/or developing or reproducing the
- * software by the user in light of its specific status of free software,
- * that may mean  that it is complicated to manipulate,  and  that  also
- * therefore means  that it is reserved for developers  and  experienced
- * professionals having in-depth computer knowledge. Users are therefore
- * encouraged to load and test the software's suitability as regards their
- * requirements in conditions enabling the security of their systems and/or
- * data to be ensured and,  more generally, to use and operate it in the
- * same conditions as regards security.
- *
- * The fact that you are presently reading this means that you have had
- * knowledge of the CeCILL license and that you accept its terms.
- *
- */
-#include "glimpse.h"
-
-int main(int argc, char *argv[])
-{
-    boost::property_tree::ptree pt;
-    po::variables_map vm;
-
-    switch (create_config(argc, argv, pt, vm)) {
-        case config_ok_go:
-            break;
-        case config_ok_halt:
-            return return_ok;
-        case config_exception:
-            return return_config_except;
-        case config_gpu_err:
-            return return_gpu_err;
-    }
-
-    return configure_and_run(pt, vm);
-
-}
-
-
-
